Monster script - Change EndTurn to FinishAction - Change EndTurn to FinishActionNoCrossFadeToIdle - Change EndTurnBehaviour to FinishActionBehaviour - Change StartCoroutineRunWaitTimeToEndTurn to StartCoroutineRunWaitTimeToFinishAction - Change RunWaitTimeToEndTurn to RunWaitTimeToFinishAction - Change parameter name turnFall to fallTime in RandomCharacterFall method

+++ Assets/Scripts/Play/Units/Character/CharacterController.cs
@@ -570,9 +570,9 @@ public class CharacterController : Unit
         }
     }
 
-    public void SetToFall(int turnFall)
+    public void SetToFall(float turnFall)
     {
         IsFall = true;
-        TurnFall = turnFall;
+        TurnFall = (int)turnFall;
     }
 }
